Source bashrc from profile and remove bash_profile
This eliminates bash_profile (which never did much anyway) and also
makes sure bashrc is sourced on login shells. If we do an exec at the
end of profile, the old bash_profile would never get around to sourcing
bashrc. Now we can ensure bashrc is sourced before we exec.